Rewrite paths in js files as well
