+++ docker/collect_static.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 set -e
 
-python manage.py rewrite_paths_in_css
+python manage.py rewrite_paths_in_static_files
 python manage.py collectstatic --noinput
